requirements.txt
discord.py
# py-cord
requests